+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file. The
 format is based on [Keep a Changelog], and this project aims to follow
 [Semantic Versioning].
 
+## [0.2.0] - 2024-05-07
+
+### Added
+
+- `MutableVecEntry` trait added
+
+### Changed
+
+- binary searching methods removed
+
 ## [0.1.9] - 2024-04-13
 
 ### Added